python/requirements.txt
flask
pycrypto
requests
